Add support for jpeg images > 64K

+++ Adafruit_VC0706.h
@@ -107,7 +107,7 @@ private:
   uint8_t serialNum;
   uint8_t camerabuff[CAMERABUFFSIZ + 1];
   uint8_t bufferLen;
-  uint16_t frameptr;
+  uint32_t frameptr;
 
 #if defined(__AVR__) || defined(ESP8266)
   SoftwareSerial *swSerial;
