+++ examples/MotionDetect/MotionDetect.ino
@@ -164,7 +164,7 @@ void loop() {
   
   File imgFile = SD.open(filename, FILE_WRITE);
   
-  uint16_t jpglen = cam.frameLength();
+  uint32_t jpglen = cam.frameLength();
   Serial.print(jpglen, DEC);
   Serial.println(" byte image");
  
@@ -173,7 +173,7 @@ void loop() {
   while (jpglen > 0) {
     // read 32 bytes at a time;
     uint8_t *buffer;
-    uint8_t bytesToRead = min((uint16_t)32, jpglen); // change 32 to 64 for a speedup but may not work with all setups!
+    uint8_t bytesToRead = min((uint32_t)32, jpglen); // change 32 to 64 for a speedup but may not work with all setups!
     buffer = cam.readPicture(bytesToRead);
     imgFile.write(buffer, bytesToRead);
